Raise exception for empty file in store dependency
* Raise exception for empty file when storing dependency

When a file download fails do to a 404, or other network issue,
`_store_dependency` will store and cache and empty file. This can be an
issue if a user has a momemtary networking or other issue. Instead this
raises and exception to avoid that situation

`RuntimeException` is likely too broad and should be something more
specific.

This is done were as opposed to the `download` method of `ExtenalData`
because that method makes use of `shell` which does not capture the
return code and does not throw and exception on error. Not changing that
functionality as it is possible there was a reason for using both
`shell` and `wget`. An alternative solution would be to make use of
`urllib` or another library that would raise and exception or make use
of `subprocess` or similar to act upon the return code.